Move vec
